requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-dotenv
openai
dvc
mlflow
optuna
pytest
kagglehub
pymupdf
tenacity
seqeval
httpx
